reorganize some more
