+++ src/photos/components/multi/GalleryPage/index.tsx
@@ -1,4 +1,4 @@
-import Modal from '@src/photos/components/single/Modal'
+import Modal from '@src/photos/components/multi/GalleryPage/Modal'
 import { getPhotoIdFromRouter } from '@src/photos/pageHelpers/getPhotoIdFromRouter'
 import escapeCloudinaryString from '@src/photos/utils/escapeCloudinaryString'
 import type { CityGallery, CityGalleryMap, CountryGallery, ImageProps } from '@src/photos/utils/types'
